refator: Rename folder(config to constants)
