+++ scripts/constant/signConfig.js
